pyproject.toml
[tool.poetry]
name = "infograph_wd_assignment_backend"
version = "0.1.0"
description = ""
authors = ["Basel Mousa Atalla <baselatalla10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
django-cors-headers = "^3.8.0"
django-environ = "^0.6.0"
psycopg2-binary = "^2.9.1"
django-rest-auth = "^0.9.5"
psycopg2 = "^2.9.1"
django-storages = "^1.11.1"
djangorestframework-simplejwt = "^4.8.0"
Pillow = "^8.3.2"
django-versatileimagefield = "^2.2"
drf-flex-fields = "^0.9.3"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
